requirements.txt
natsort==7.1.1
numpy==1.20.1
pandas==1.2.3
PyYAML==6.0
